backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic-ai==0.0.14
python-dotenv==1.0.0
httpx>=0.27.2
pydantic>=2.9.0
griffe==1.5.0
pymongo[srv]>=4.9.0,<4.10
motor==3.6.0
bcrypt==4.2.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
email-validator>=2.0.0
certifi>=2024.0.0
aiosmtplib==3.0.1
